[UDPDataReceiveSequence][BUGFIX] function won't accept static method pointers directly
Issue details:
    std::function won't accept static method pointers directly
    The ternary result needs both sides to be the exact same type — static methods decay to function pointers, not std::function

Because of this issue. Max memory calculated is 0, so UDP data was not accepting.
Fixed this issue

+++ app/code/include/DataPool/DataPool.hpp
@@ -17,8 +17,8 @@ namespace sensormoniteringhub{
         /// @brief store config limit always in bytes internally and convert at the boundary (when reading config)
         struct MemoryConfig {
             size_t maxBytes;
-            static size_t FromKB(uint16_t kb) { return kb * 1024; }
-            static size_t FromMB(uint16_t mb) { return mb * 1024 * 1024; }
+            static size_t FromKB(uint16_t kb) { return static_cast<size_t>(kb) * 1024; }
+            static size_t FromMB(uint16_t mb) { return static_cast<size_t>(mb) * 1024 * 1024; }
         };
         class DataPool : public IEvents{
             private:
